requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0

# Audio Processing
openai-whisper==20231117
pydub==0.25.1

# Embeddings and ML
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1

# Vector Database
faiss-cpu==1.7.4

# LLM
llama-cpp-python==0.2.20

# Utilities
numpy==1.24.3
pandas==2.1.3
aiofiles==23.2.1
